requirements.txt
access==1.1.1
affine==2.3.0
alabaster==0.7.12
alembic==1.4.2
aniso8601==8.0.0
appdirs==1.4.4
astroid==2.3.3
attrs==20.2.0
Babel==2.8.0
backcall==0.2.0
bandit==1.6.2
beautifulsoup4==4.8.2
bs4==0.0.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.7.0
coverage==5.2.1
cycler==0.10.0
dataset==1.3.2
dateparser==0.7.6
decorator==4.4.2
descartes==1.1.0
distlib==0.3.1
docutils==0.16
esda==2.3.1
filelock==3.0.12
Fiona==1.8.17
flake8==3.8.3
Flask==1.1.2
Flask-RESTful==0.3.8
future==0.18.2
geocoder==1.38.1
geographiclib==1.50
geopandas==0.8.1
geopy==2.0.0
giddy==2.3.3
gitdb==4.0.5
GitPython==3.1.11
google-api-core==1.22.0
google-auth==1.19.2
google-cloud-core==1.3.0
google-cloud-storage==1.29.0
google-cloud-translate==2.0.1
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
grpcio==1.30.0
html5lib==1.1
idna==2.10
imagesize==1.2.0
inequality==1.0.0
iniconfig==1.1.1
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
jupyter-client==6.1.7
jupyter-core==4.6.3
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
libpysal==4.3.0
llvmlite==0.34.0
lxml==4.5.0
Mako==1.1.3
mapclassify==2.3.0
MarkupSafe==1.1.1
matplotlib==3.3.1
mccabe==0.6.1
mgwr==2.1.2
more-itertools==8.6.0
mpmath==1.1.0
munch==2.5.0
mypy==0.782
mypy-extensions==0.4.3
networkx==2.5
numba==0.51.2
numpy==1.18.2
opencv-contrib-python==4.4.0.46
packaging==20.4
pandas==1.1.1
parso==0.7.1
patsy==0.5.1
pbr==5.5.1
pexpect==4.8.0
phantomjs==1.3.0
pickleshare==0.7.5
Pillow==7.2.0
pluggy==0.13.1
pointpats==2.2.0
prompt-toolkit==3.0.8
protobuf==3.12.2
psycopg2==2.8.6
psycopg2-binary==2.8.4
ptyprocess==0.6.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pydataset==0.2.0
pyflakes==2.2.0
Pygments==2.7.2
pylint==2.4.4
pymongo==3.11.0
pyparsing==2.4.7
pyproj==2.6.1.post1
pysal==2.3.0
pytest==6.0.1
pytest-cov==2.10.1
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.2
quantecon==0.4.8
rasterio==1.1.8
rasterstats==0.15.0
ratelim==0.1.6
regex==2020.7.14
requests==2.24.0
rsa==4.6
Rtree==0.9.4
scikit-learn==0.23.2
scipy==1.5.4
seaborn==0.11.0
segregation==1.4.0
selenium==3.141.0
Shapely==1.7.1
sidrapy==0.1.1
simplejson==3.17.2
six==1.15.0
smmap==3.0.4
snowballstemmer==2.0.0
snuggs==1.4.7
soupsieve==2.0
spaghetti==1.5.3
spglm==1.0.8
Sphinx==2.4.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
spint==1.0.7
splot==1.1.3
spreg==1.1.2.post1
spvcm==0.3.0
SQLAlchemy==1.3.12
statsmodels==0.12.1
stevedore==3.2.2
sympy==1.6.2
threadpoolctl==2.1.0
tobler==0.4.0
toml==0.10.1
tornado==6.1
tqdm==4.51.0
traitlets==5.0.5
typed-ast==1.4.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.24.3
virtualenv==20.1.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.11.2
